backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Email
python-dotenv==1.0.0
aiosmtplib==3.0.1
email-validator==2.1.0

# CORS
python-cors==1.0.0

# Image handling
Pillow==10.2.0

# Utilities
python-slugify==8.0.1
